Version 3
Fixed inside of the warehouse not being rendered when coming up from the sewers.
Adjusted the height of the ladders leading down to the sewer.
Reduced the number of floating objects.
Reworked top safehouse bathroom to make hiding bodies in it easier.
Adjusted the position of the civilian at administration.
Fixed some lights not turning off during the blackout.
Fixed server racks not accepting the portable server after any other loot was thrown infront of it.